+++ generator/src/main/java/com/scottlogic/deg/generator/validators/GenerationConfigValidator.java
@@ -1,4 +1,4 @@
-package com.scottlogic.deg.generator.CommandLine;
+package com.scottlogic.deg.generator.validators;
 
 import com.scottlogic.deg.generator.generation.GenerationConfig;
